set emoji subcommand
