--- src/commands/projects/SetEmoji.ts
@@ -1,38 +1,0 @@
-import { ChatInputCommandInteraction, PermissionFlagsBits, SlashCommandBuilder } from "discord.js";
-import CrossingGuardBot from "../../CrossingGuardBot";
-
-const data = new SlashCommandBuilder()
-    .setName("setemoji")
-    .setDescription("Set a project's emoji")
-    .setDefaultMemberPermissions(PermissionFlagsBits.Administrator)
-    .addStringOption(option =>
-        option.setName("project_name")
-            .setDescription("The name of the project")
-            .setRequired(true))
-    .addStringOption(option =>
-        option.setName("emoji_string")
-            .setDescription("The id of this emoji if custom, otherwise the unicode character")
-            .setRequired(true));
-
-async function execute(interaction: ChatInputCommandInteraction) {
-    const projectName = interaction.options.getString("project_name");
-    const emojiIdOrUnicode = interaction.options.getString("emoji_string");
-
-    if (!projectName || !emojiIdOrUnicode)
-        return;
-
-    CrossingGuardBot.getInstance().database.getProjectByName(projectName).then(project => {
-        if (!project) {
-            interaction.reply({ content: `No project matched the name ${projectName}`, ephemeral: true });
-            return;
-        }
-
-        project.emoji = emojiIdOrUnicode;
-        CrossingGuardBot.getInstance().database.saveProject(project);
-        interaction.reply({ content: `${project.displayName}'s emoji set to \`${project.emoji}\``, ephemeral: true });
-    });
-
-}
-
-export { data, execute };
-
